requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
dj-database-url==0.5.0
Django==3.1.6
django-environ==0.4.5
djangorestframework==3.12.2
gunicorn==20.0.4
httpie==2.4.0
idna==2.10
Pillow==8.0.1
psycopg2==2.8.6
Pygments==2.8.0
PyJWT==2.0.1
PySocks==1.7.1
python-decouple==3.4
pytz==2021.1
requests==2.25.1
requests-toolbelt==0.9.1
sqlparse==0.4.1
urllib3==1.26.3
whitenoise==5.2.0
Pillow==8.0.1
psycopg2-binary
4eabaececc20edf145355e65a3be18ac030f3de4
